packetvwr/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy.orm
databases[aiosqlite]
python-dotenv
rich
asyncpg
psycopg2-binary
cassandra-driver
simplejson


